Refactored discussion structured to use Link to change discussion to improve accessibility
